backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
psycopg2-binary
yfinance
pandas
requests
python-dotenv
pydantic
